+++ src/main/java/com/muran/dto/ActivityInfo.java
@@ -25,13 +25,13 @@ public class ActivityInfo   {
   private Date startTime = null;
   private Date endTime = null;
   private String location = null;
-  private Long signNum = null;
+  private BigInteger signNum = null;
   private Date publishTime = null;
   private String publishMan = null;
   private String content = null;
-  private Long signupTop = null;
-  private Long commentNum = null;
-  private Long status = null;
+  private BigInteger signupTop = null;
+  private BigInteger commentNum = null;
+  private BigInteger status = null;
   private Date signupEndTime = null;
 
   /**
@@ -155,7 +155,7 @@ public class ActivityInfo   {
 
   /**
    **/
-  public ActivityInfo signNum(Long signNum) {
+  public ActivityInfo signNum(BigInteger signNum) {
     this.signNum = signNum;
     return this;
   }
@@ -163,10 +163,10 @@ public class ActivityInfo   {
   
   @ApiModelProperty(value = "")
   @JsonProperty("signNum")
-  public Long getSignNum() {
+  public BigInteger getSignNum() {
     return signNum;
   }
-  public void setSignNum(Long signNum) {
+  public void setSignNum(BigInteger signNum) {
     this.signNum = signNum;
   }
 
@@ -223,7 +223,7 @@ public class ActivityInfo   {
 
   /**
    **/
-  public ActivityInfo signupTop(Long signupTop) {
+  public ActivityInfo signupTop(BigInteger signupTop) {
     this.signupTop = signupTop;
     return this;
   }
@@ -231,16 +231,16 @@ public class ActivityInfo   {
   
   @ApiModelProperty(value = "")
   @JsonProperty("signupTop")
-  public Long getSignupTop() {
+  public BigInteger getSignupTop() {
     return signupTop;
   }
-  public void setSignupTop(Long signupTop) {
+  public void setSignupTop(BigInteger signupTop) {
     this.signupTop = signupTop;
   }
 
   /**
    **/
-  public ActivityInfo commentNum(Long commentNum) {
+  public ActivityInfo commentNum(BigInteger commentNum) {
     this.commentNum = commentNum;
     return this;
   }
@@ -248,16 +248,16 @@ public class ActivityInfo   {
   
   @ApiModelProperty(value = "")
   @JsonProperty("commentNum")
-  public Long getCommentNum() {
+  public BigInteger getCommentNum() {
     return commentNum;
   }
-  public void setCommentNum(Long commentNum) {
+  public void setCommentNum(BigInteger commentNum) {
     this.commentNum = commentNum;
   }
 
   /**
    **/
-  public ActivityInfo status(Long status) {
+  public ActivityInfo status(BigInteger status) {
     this.status = status;
     return this;
   }
@@ -265,10 +265,10 @@ public class ActivityInfo   {
   
   @ApiModelProperty(value = "")
   @JsonProperty("status")
-  public Long getStatus() {
+  public BigInteger getStatus() {
     return status;
   }
-  public void setStatus(Long status) {
+  public void setStatus(BigInteger status) {
     this.status = status;
   }
